minor bug fixes to get test script running

+++ benchmarks/scripts/stages/testStages/stage2.sh
@@ -6,6 +6,8 @@ set -x
 
 echo "This is stage2"
 
+[ -f artifact1.txt ] || exit 1
+
 mkdir foo1
 echo "artifact2" > foo1/artifact2.txt
 echo "artifact3" > foo1/artifact3.txt
